Include detailed Git info in Open Dylan version string
This uses the same mechanism as Deft to maintain version info, so `dylan-compiler
-version` displays something similar to "v2025.1.0-13-g7f7fdb9 built on
2025-01-08T21:05:44-05:00".

Besides more detail in the version, making it easier to pinpoint the commit from which
Open Dylan was built, this also makes it so that we don't have to manually update the
version string.

The release tag must be applied before the build, which is a minor change in how we've
traditionally done releases.