doc: Add dark mode for diagrams system:file-system
I have not found a way to use the CSS class 'only-light' and
'only-dark' with the Graphviz inline diagrams. My solution has been
create the .png from the graphviz source and from that, the dark
version.

To allow the caption, I use an adminition and inside it, the image
(figure allows captions, but image does not).

Closes #1714