system: Improve resolve-locator and directory-contents
directory-contents returns a `<file-locator>` even for a symlink that resolves to a
directory. Add a `resolve-links?` keyword to resolve links. This uncovered a bug in
resolve-locator such that it would always return a locator of the same class that it was
passed, regardless of the file system entity it resolved to.