dfmc-optimization: Add types to lifted references
Though lambda lifting presently is done late within the optimization
phase, any re-optimization that occurs will result in losing the
original type estimates for the lifted references. Since the optimizer
presently does not do interprocedural type estimation, the best
approach is to record the types of the lifted reference as the
new parameter types.

* sources/dfmc/optimization/closure.dylan
  (lift-reference): Estimate the type of closed-over references
   and pass it to add-function-argument.
  (add-function-argument): Receive the type of the new parameter
   and include it (rather than the original <object>) in the
   function signature.
  (add-call-argument): Mark the augmented call for re-optimization.