release-info: Remove information about merged DLLs
The original Functional Developer build system was capable of
combining multiple libraries into a merged DLL for final
deployment. Unfortunately the only way of declaring that libraries
should be merged was via Dylan Library Pack specification files,
meaning that this potentially useful facility was not available to
ordinary user projects. Given this, the new build system never
implemented library merging, and most of the consumers of this
information were removed.

This commit removes the remaining vestigal code tracking merged DLLs.

* sources/lib/release-info/libraries.dylan
  (info-merge-parent): Remove slot from <library-info>.
  (info-merged-libraries): Remove slot from <library-binary-info>.
  (interpret-library-binary-xml): Remove computation of the merged
   libraries that make up a DLL.
  (merged-project-name, merged-project-libraries): Remove.

* sources/lib/release-info/module.dylan
  (module release-info): Remove export of info-merged-libraries,
  info-merge-parent, merged-project-name, and
  merged-project-libraries.

* sources/dfmc/execution/runtime-execution.dylan
  (load-dll): Remove lookup for the merged library case.

* sources/environment/dfmc/database/library-objects.dylan
  (project-executable-pathname): Remove lookup for the merged library
   case.

* sources/environment/commands/library-packs.dylan
  (describe-state on <library-info>): Remove lookup and printing
   of merged libraries.